Create Subset Product.cpp
You are given an array arr if integers of length n. your task is to check whether it is possible to divide the array into 2 parts such that the first part contains one of the elements of the array and the second part contains all other elements, and the product of all the elements in the second part equal the value of the element in the first part. If it is possible to achieve this, return the index (0-indexed) of the element you have chosen as your first part. If there are multiple answers, return the answer with highest index. If is not possible to divide the array in this way, return -1.

Input: arr = [2, 8, 4, 1], n = 4
Output: 1
Explanation : We see that 8 = 2 * 4 *1, and 1 is the index of 8

Input: arr = [1,5,7,9,3], n =5
Output: -1

Input: arr = [4, 1, 4 ], n =3
Output: 2